+++ packages/entities/entities-routes/src/components/RouteList.vue
@@ -14,6 +14,7 @@
       :query="filterQuery"
       :table-headers="tableHeaders"
       :title="title"
+      :title-tag="titleTag"
       :use-action-outside="useActionOutside"
       @clear-search-input="clearFilter"
       @click:row="(row: any) => rowClick(row as EntityRow)"
@@ -185,7 +186,7 @@ import type { AxiosError } from 'axios'
 import { useRouter } from 'vue-router'
 
 import { BadgeMethodAppearances } from '@kong/kongponents'
-import type { BadgeMethodAppearance } from '@kong/kongponents'
+import type { BadgeMethodAppearance, HeaderTag } from '@kong/kongponents'
 import {
   EntityBaseTable,
   EntityDeleteModal,
@@ -268,6 +269,10 @@ const props = defineProps({
   title: {
     type: String,
     default: '',
+  },
+  titleTag: {
+    type: String as PropType<HeaderTag>,
+    default: 'h2',
   },
   /** default to false, setting to true will teleport the toolbar button to the destination in the consuming app */
   useActionOutside: {