+++ packages/entities/entities-shared/src/components/entity-base-table/EntityBaseTable.vue
@@ -2,6 +2,7 @@
   <KCard
     class="kong-ui-entity-base-table"
     :title="title"
+    :title-tag="titleTag"
   >
     <KTable
       ref="tableRefs"
@@ -114,7 +115,7 @@ import { computed, ref } from 'vue'
 import composables from '../../composables'
 import { useTablePreferences } from '@kong-ui-public/core'
 import type { UserTablePreferences } from '@kong-ui-public/core'
-import type { SwrvStateData } from '@kong/kongponents'
+import type { SwrvStateData, HeaderTag } from '@kong/kongponents'
 import { KUI_COLOR_TEXT_NEUTRAL_STRONGER } from '@kong/design-tokens'
 import EntityBaseTableCell from './EntityBaseTableCell.vue'
 
@@ -228,6 +229,10 @@ const props = defineProps({
   title: {
     type: String,
     default: '',
+  },
+  titleTag: {
+    type: String as PropType<HeaderTag>,
+    default: 'h2',
   },
   /** default to false, setting to true will suppress the row click event even if "@click:row" is attached */
   disableRowClick: {