+++ packages/analytics/analytics-chart/CHANGELOG.md
@@ -3,6 +3,54 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.0.1](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/analytics-chart@3.0.0...@kong-ui-public/analytics-chart@3.0.1) (2024-03-25)
+
+
+### Bug Fixes
+
+* **entities-plugins, vfg:** create plugin form a11y improvements [KHCP-11026] ([#1272](https://github.com/Kong/public-ui-components/issues/1272)) ([b1ed04e](https://github.com/Kong/public-ui-components/commit/b1ed04eb56ff2869b988a217220debc78655f3f0))
+
+
+
+
+
+# [3.0.0](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/analytics-chart@2.5.5...@kong-ui-public/analytics-chart@3.0.0) (2024-03-21)
+
+
+### Bug Fixes
+
+* **analytics-chart:** remove fill prop [MA-2659] ([#1280](https://github.com/Kong/public-ui-components/issues/1280)) ([b24f07d](https://github.com/Kong/public-ui-components/commit/b24f07d6e230116c619e690b35d030cfbfb37bd9))
+
+
+### BREAKING CHANGES
+
+* **analytics-chart:** removed the fill prop from chart options
+
+* fix: no need to pass fill to chart options anymore in dashboard renderer
+
+
+
+
+
+## [2.5.5](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/analytics-chart@2.5.4...@kong-ui-public/analytics-chart@2.5.5) (2024-03-21)
+
+**Note:** Version bump only for package @kong-ui-public/analytics-chart
+
+
+
+
+
+## [2.5.4](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/analytics-chart@2.5.3...@kong-ui-public/analytics-chart@2.5.4) (2024-03-20)
+
+
+### Bug Fixes
+
+* proper heading element in widgets [KHCP-11027] ([#1252](https://github.com/Kong/public-ui-components/issues/1252)) ([b66ce71](https://github.com/Kong/public-ui-components/commit/b66ce711fa876bb75bb880eece8514a8055034be))
+
+
+
+
+
 ## [2.5.3](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/analytics-chart@2.5.2...@kong-ui-public/analytics-chart@2.5.3) (2024-03-20)